backend/requirements.txt
# Slack Bot Dependencies
slack-bolt==1.18.0
python-dotenv==1.0.0

# Database
supabase==2.3.0

# AI/LLM
openai==1.12.0

# Async Support
aiohttp==3.9.1
